+++ Hearthstone Deck Tracker/Windows/OpponentWindow.xaml.cs
@@ -54,8 +54,6 @@ namespace Hearthstone_Deck_Tracker
 
 		public List<Card> OpponentDeck => _game.Opponent.OpponentCardList;
 
-		public bool ShowToolTip => Config.Instance.WindowCardToolTips;
-
 		public event PropertyChangedEventHandler? PropertyChanged;
 
 		public void Update()